ID: Integer -> Long (better practice) + bigint in flyway  / List -> Set (uniqueness) / AssignmentLogic -> extracting it, hence reducing duplicate code / Pagination to async operation result / More explaining comments / Better logic structure / TypeMismatchException -> exception handling

+++ src/main/java/org/example/driverandfleetmanagementapp/dto/DriverBasicDto.java
@@ -12,8 +12,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class DriverBasicDto {
-    private Integer id;
+
+    private Long id;
     private String firstName;
     private String lastName;
     private String licenseNumber;
+
 }
